+++ sdk/compute/arm-compute/src/operations/capacityReservationGroups.ts
@@ -174,8 +174,8 @@ export class CapacityReservationGroupsImpl
 
   /**
    * The operation to create or update a capacity reservation group. When updating a capacity reservation
-   * group, only tags may be modified. Please refer to https://aka.ms/CapacityReservation for more
-   * details.
+   * group, only tags and sharing profile may be modified. Please refer to
+   * https://aka.ms/CapacityReservation for more details.
    * @param resourceGroupName The name of the resource group.
    * @param capacityReservationGroupName The name of the capacity reservation group.
    * @param parameters Parameters supplied to the Create capacity reservation Group.
@@ -195,7 +195,7 @@ export class CapacityReservationGroupsImpl
 
   /**
    * The operation to update a capacity reservation group. When updating a capacity reservation group,
-   * only tags may be modified.
+   * only tags and sharing profile may be modified.
    * @param resourceGroupName The name of the resource group.
    * @param capacityReservationGroupName The name of the capacity reservation group.
    * @param parameters Parameters supplied to the Update capacity reservation Group operation.